+++ src/config.php
@@ -15,21 +15,20 @@ return [
 
     // Store tags locally and purge Urls
     // In case the cache driver does not support tag purging
-    // MySql only so far
-    'useLocalTags' => true,
+    'useLocalTags'  => true,
 
     // Drivers settings
     'drivers'       => [
 
         // Varnish config
-        'varnish' => [
+        'varnish'    => [
             'tagHeaderName'   => 'XKEY',
             'purgeHeaderName' => 'XKEY-PURGE',
-            'purgeUrl'        => getenv('VARNISH_URL') ?? 'http://127.0.0.1:80/',
+            'purgeUrl'        => getenv('VARNISH_URL') ?: 'http://127.0.0.1:80/',
         ],
 
         // Fastly config
-        'fastly'  => [
+        'fastly'     => [
             'tagHeaderName' => 'Surrogate-Key',
             'serviceId'     => getenv('FASTLY_SERVICE_ID'),
             'apiToken'      => getenv('FASTLY_API_TOKEN'),
@@ -37,15 +36,25 @@ return [
         ],
 
         // KeyCDN config
-        'keycdn'  => [
+        'keycdn'     => [
             'tagHeaderName' => 'Cache-Tag',
             'apiKey'        => getenv('KEYCDN_API_KEY'),
             'zoneId'        => getenv('KEYCDN_ZONE_ID'),
             'zoneUrl'       => getenv('KEYCDN_ZONE_URL')
         ],
 
+        // CloudFlare config
+        'cloudflare' => [
+            'tagHeaderName'      => 'Cache-Tag',
+            'tagHeaderDelimiter' => ',',
+            'apiKey'             => getenv('CLOUDFLARE_API_KEY'),
+            'apiEmail'           => getenv('CLOUDFLARE_API_EMAIL'),
+            'zoneId'             => getenv('CLOUDFLARE_ZONE_ID'),
+            'domain'             => getenv('CLOUDFLARE_DOMAIN')
+        ],
+
         // Dummy driver (default)
-        'dummy'   => [
+        'dummy'      => [
             'tagHeaderName'   => 'X-CacheTag',
             'logPurgeActions' => true,
         ]